[refactor] #133 pasteboard Live, Test 분리
