+++ Projects/CoreKit/Sources/Data/Client/PasteboardClient/PasteboardClient.swift
@@ -1,21 +1,15 @@
 //
-//  Interface.swift
-//  App
+//  PasteboardClient.swift
+//  CoreKit
 //
 //  Created by 김민호 on 8/2/24.
 //
 
 import Foundation
-import Dependencies
+import DependenciesMacros
 
-extension DependencyValues {
-  public var pasteboard: PasteboardClient {
-    get { self[PasteboardClient.self] }
-    set { self[PasteboardClient.self] = newValue }
-  }
-}
-
+@DependencyClient
 public struct PasteboardClient: Sendable {
-  public var changes: @Sendable () -> AsyncStream<Void>
-  public var probableWebURL: @Sendable () async throws -> URL?
+    public var changes: @Sendable () -> AsyncStream<Void> = { .finished }
+    public var probableWebURL: @Sendable () async throws -> URL?
 }
